add-message="添加日期拦截"

+++ OAuth-ResourceServer/src/main/java/com/wojucai/dao/ClientRepository.java
@@ -17,7 +17,7 @@ public interface ClientRepository extends JpaRepository<Client, Integer> {
      * @param id 客户端id
      * @return 返回是否删除成功
      */
-    int deleteByClientId(Integer id);
+    int deleteByClientId(Long id);
 
 
 }
